Crate risc0_binfmt

Manages formatted binaries used by the RISC Zero zkVM

Structs

InvalidExitCodeError

Error returned when a (system, user) exit code pair is an invalid representation.

MemoryImage

An image of a zkVM guest’s memory

PageTableInfo

Structure representing the page table for zkVM memory.

Program

A RISC Zero program

SystemState

Represents the public state of a segment, needed for continuations and receipt verification.

Enums

DecodeError

Error returned when a decoding failure occurs.

ExitCode

Exit condition indicated by the zkVM at the end of the guest execution.

Traits

Digestible

Defines a collision resistant hash for the typed and structured data.

Functions

compute_image_id

Compute and return the ImageID of the specified ELF binary.

read_sha_halfs

Read a SHA-256 digest as a series of half-wrods (i.e. 16-bit values).

tagged_iter

A list hashing routine, permitting iterative opening over elements.

tagged_list

A list hashing routine, permitting iterative opening over elements.

tagged_list_cons

Calculate the hash resulting from adding one element to a tagged_list digest.

tagged_struct

A struct hashing routine, permitting tree-like opening of fields.

write_sha_halfs

Write a SHA-256 digest as a series of half-wrods (i.e. 16-bit values).