risc0_binfmt
Function tagged_list

pub fn tagged_list<S: Sha256>(tag: &str, list: &[impl Borrow<Digest>]) -> Digest

A list hashing routine, permitting iterative opening over elements.

Used for hashing of the receipt claim assumptions list, and in the recursion predicates.